+++ llama_index/response_synthesizers/refine.py
@@ -1,9 +1,10 @@
 import logging
-from typing import Any, Generator, Optional, Sequence, cast
-
+from typing import Any, Generator, Optional, Sequence, cast, Type, Callable
+from pydantic import BaseModel, Field
 from llama_index.indices.service_context import ServiceContext
+from llama_index.prompts.base import BasePromptTemplate, PromptTemplate
+from llama_index.llm_predictor.base import BaseLLMPredictor
 from llama_index.indices.utils import truncate_text
-from llama_index.prompts import BasePromptTemplate
 from llama_index.prompts.default_prompt_selectors import (
     DEFAULT_REFINE_PROMPT_SEL,
     DEFAULT_TEXT_QA_PROMPT_SEL,
@@ -11,8 +12,58 @@ from llama_index.prompts.default_prompt_selectors import (
 from llama_index.response.utils import get_response_text
 from llama_index.response_synthesizers.base import BaseSynthesizer
 from llama_index.types import RESPONSE_TEXT_TYPE
+from llama_index.program.base_program import BasePydanticProgram
+from llama_index.program.openai_program import OpenAIPydanticProgram
+from llama_index.program.llm_program import LLMTextCompletionProgram
+from llama_index.output_parsers.pydantic import PydanticOutputParser
 
 logger = logging.getLogger(__name__)
+
+
+class StructuredRefineResponse(BaseModel):
+    """
+    Used to answer a given query based on the provided context.
+
+    Also indicates if the query was satisfied with the provided answer.
+    """
+
+    answer: str = Field(
+        description="The answer for the given query, based on the context and not "
+        "prior knowledge."
+    )
+    query_satisfied: bool = Field(
+        description="True if there was enough context given to provide an answer "
+        "that satisfies the query."
+    )
+
+
+class DefaultRefineProgram(BasePydanticProgram):
+    """
+    Runs the query on the LLM as normal and always returns the answer with
+    query_satisfied=True. In effect, doesn't do any answer filtering.
+    """
+
+    def __init__(self, prompt: BasePromptTemplate, llm_predictor: BaseLLMPredictor):
+        self._prompt = prompt
+        self._llm_predictor = llm_predictor
+
+    @property
+    def output_cls(self) -> Type[BaseModel]:
+        return StructuredRefineResponse
+
+    def __call__(self, *args: Any, **kwds: Any) -> StructuredRefineResponse:
+        answer = self._llm_predictor.predict(
+            self._prompt,
+            **kwds,
+        )
+        return StructuredRefineResponse(answer=answer, query_satisfied=True)
+
+    async def acall(self, *args: Any, **kwds: Any) -> StructuredRefineResponse:
+        answer = await self._llm_predictor.apredict(
+            self._prompt,
+            **kwds,
+        )
+        return StructuredRefineResponse(answer=answer, query_satisfied=True)
 
 
 class Refine(BaseSynthesizer):
@@ -25,11 +76,26 @@ class Refine(BaseSynthesizer):
         refine_template: Optional[BasePromptTemplate] = None,
         streaming: bool = False,
         verbose: bool = False,
+        structured_answer_filtering: bool = False,
+        program_factory: Optional[
+            Callable[[BasePromptTemplate], BasePydanticProgram]
+        ] = None,
     ) -> None:
         super().__init__(service_context=service_context, streaming=streaming)
         self._text_qa_template = text_qa_template or DEFAULT_TEXT_QA_PROMPT_SEL
         self._refine_template = refine_template or DEFAULT_REFINE_PROMPT_SEL
         self._verbose = verbose
+        self._structured_answer_filtering = structured_answer_filtering
+
+        if self._streaming and self._structured_answer_filtering:
+            raise ValueError(
+                "Streaming not supported with structured answer filtering."
+            )
+        if not self._structured_answer_filtering and program_factory is not None:
+            raise ValueError(
+                "Program factory not supported without structured answer filtering."
+            )
+        self._program_factory = program_factory or self._default_program_factory
 
     def get_response(
         self,
@@ -61,6 +127,29 @@ class Refine(BaseSynthesizer):
             response = cast(Generator, response)
         return response
 
+    def _default_program_factory(self, prompt: PromptTemplate) -> BasePydanticProgram:
+        if self._structured_answer_filtering:
+            try:
+                return OpenAIPydanticProgram.from_defaults(
+                    StructuredRefineResponse,
+                    prompt=prompt,
+                    llm=self._service_context.llm,
+                    verbose=self._verbose,
+                )
+            except ValueError:
+                output_parser = PydanticOutputParser(StructuredRefineResponse)
+                return LLMTextCompletionProgram.from_defaults(
+                    output_parser,
+                    prompt=prompt,
+                    llm=self._service_context.llm,
+                    verbose=self._verbose,
+                )
+        else:
+            return DefaultRefineProgram(
+                prompt=prompt,
+                llm_predictor=self._service_context.llm_predictor,
+            )
+
     def _give_response_single(
         self,
         query_str: str,
@@ -74,24 +163,31 @@ class Refine(BaseSynthesizer):
         )
 
         response: Optional[RESPONSE_TEXT_TYPE] = None
+        program = self._program_factory(text_qa_template)
         # TODO: consolidate with loop in get_response_default
         for cur_text_chunk in text_chunks:
+            query_satisfied = False
             if response is None and not self._streaming:
-                response = self._service_context.llm_predictor.predict(
-                    text_qa_template,
-                    context_str=cur_text_chunk,
+                structured_response = cast(
+                    StructuredRefineResponse, program(context_str=cur_text_chunk)
                 )
+                query_satisfied = structured_response.query_satisfied
+                if query_satisfied:
+                    response = structured_response.answer
             elif response is None and self._streaming:
                 response = self._service_context.llm_predictor.stream(
                     text_qa_template,
                     context_str=cur_text_chunk,
                 )
+                query_satisfied = True
             else:
                 response = self._refine_response_single(
                     cast(RESPONSE_TEXT_TYPE, response),
                     query_str,
                     cur_text_chunk,
                 )
+        if response is None:
+            response = "Empty Response"
         if isinstance(response, str):
             response = response or "Empty Response"
         else:
@@ -104,7 +200,7 @@ class Refine(BaseSynthesizer):
         query_str: str,
         text_chunk: str,
         **response_kwargs: Any,
-    ) -> RESPONSE_TEXT_TYPE:
+    ) -> Optional[RESPONSE_TEXT_TYPE]:
         """Refine response."""
         # TODO: consolidate with logic in response/schema.py
         if isinstance(response, Generator):
@@ -122,20 +218,26 @@ class Refine(BaseSynthesizer):
             refine_template, text_chunks=[text_chunk]
         )
 
+        program = self._program_factory(refine_template)
         for cur_text_chunk in text_chunks:
+            query_satisfied = False
             if not self._streaming:
-                response = self._service_context.llm_predictor.predict(
-                    refine_template,
-                    context_msg=cur_text_chunk,
+                structured_response = cast(
+                    StructuredRefineResponse, program(context_msg=cur_text_chunk)
                 )
+                query_satisfied = structured_response.query_satisfied
+                if query_satisfied:
+                    response = structured_response.answer
             else:
                 response = self._service_context.llm_predictor.stream(
                     refine_template,
                     context_msg=cur_text_chunk,
                 )
-            refine_template = self._refine_template.partial_format(
-                query_str=query_str, existing_answer=response
-            )
+                query_satisfied = True
+            if query_satisfied:
+                refine_template = self._refine_template.partial_format(
+                    query_str=query_str, existing_answer=response
+                )
 
         return response
 
@@ -162,6 +264,8 @@ class Refine(BaseSynthesizer):
                     prev_response_obj, query_str, text_chunk
                 )
             prev_response_obj = response
+        if response is None:
+            response = "Empty Response"
         if isinstance(response, str):
             response = response or "Empty Response"
         else:
@@ -174,7 +278,7 @@ class Refine(BaseSynthesizer):
         query_str: str,
         text_chunk: str,
         **response_kwargs: Any,
-    ) -> RESPONSE_TEXT_TYPE:
+    ) -> Optional[RESPONSE_TEXT_TYPE]:
         """Refine response."""
         # TODO: consolidate with logic in response/schema.py
         if isinstance(response, Generator):
@@ -190,18 +294,24 @@ class Refine(BaseSynthesizer):
             refine_template, text_chunks=[text_chunk]
         )
 
+        program = self._program_factory(refine_template)
         for cur_text_chunk in text_chunks:
+            query_satisfied = False
             if not self._streaming:
-                response = await self._service_context.llm_predictor.apredict(
-                    refine_template,
-                    context_msg=cur_text_chunk,
+                structured_response = await program.acall(context_msg=cur_text_chunk)
+                structured_response = cast(
+                    StructuredRefineResponse, structured_response
                 )
+                query_satisfied = structured_response.query_satisfied
+                if query_satisfied:
+                    response = structured_response.answer
             else:
                 raise ValueError("Streaming not supported for async")
 
-            refine_template = self._refine_template.partial_format(
-                query_str=query_str, existing_answer=response
-            )
+            if query_satisfied:
+                refine_template = self._refine_template.partial_format(
+                    query_str=query_str, existing_answer=response
+                )
 
         return response
 
@@ -218,13 +328,17 @@ class Refine(BaseSynthesizer):
         )
 
         response: Optional[RESPONSE_TEXT_TYPE] = None
+        program = self._program_factory(text_qa_template)
         # TODO: consolidate with loop in get_response_default
         for cur_text_chunk in text_chunks:
             if response is None and not self._streaming:
-                response = await self._service_context.llm_predictor.apredict(
-                    text_qa_template,
-                    context_str=cur_text_chunk,
+                structured_response = await program.acall(context_str=cur_text_chunk)
+                structured_response = cast(
+                    StructuredRefineResponse, structured_response
                 )
+                query_satisfied = structured_response.query_satisfied
+                if query_satisfied:
+                    response = structured_response.answer
             elif response is None and self._streaming:
                 raise ValueError("Streaming not supported for async")
             else:
@@ -233,6 +347,8 @@ class Refine(BaseSynthesizer):
                     query_str,
                     cur_text_chunk,
                 )
+        if response is None:
+            response = "Empty Response"
         if isinstance(response, str):
             response = response or "Empty Response"
         else: