+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added support for `MonsterLLM` using MonsterAPI (#7343) 
 - Support comments fields in NebulaGraphStore and int type VID (#7402)
 - Added configurable endpoint for DynamoDB (#6777)
+- Add structured answer filtering for Refine response synthesizer (#7317)
 
 ### Bug Fixes / Nits
 - Use `utf-8` for json file reader (#7390)